Remove noSelection panel.
