+++ assets/src/edit-story/components/panels/design/index.js
@@ -26,7 +26,6 @@ export { default as ImageAccessibilityPanel } from './imageAccessibility';
 export { default as LayerPanel } from './layer';
 export { default as LayerStylePanel } from './layerStyle';
 export { default as LinkPanel } from './link';
-export { default as NoSelectionPanel } from './noSelection';
 export { default as PageAttachmentPanel } from './pageAttachment';
 export { default as PageStylePanel } from './pageStyle';
 export { default as ShapeStylePanel } from './shapeStyle';
